requirements.txt
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
anyio==4.5.2
attrs==23.2.0
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==3.31.2
contourpy==1.1.1
cycler==0.12.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.16.1
Flask==2.3.3
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.0
h11==0.14.0
h5py==3.11.0
httptools==0.6.4
huggingface-hub==0.30.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openslide-bin==4.0.0.6
openslide-python==1.4.1
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
redis==3.5.3
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
structlog==24.4.0
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tokenizers==0.20.3
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.67.1
transformers==4.46.3
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
zipp==3.20.2
